"[ERROR] MinecartMeter/src/main/java/no/openshell/oddstr13/minecartmeter/MinecartMeterListener.java:[46,61] possible loss of precision" oh, stfu

+++ src/main/java/no/openshell/oddstr13/minecartmeter/MinecartMeterListener.java
@@ -42,8 +42,8 @@ public class MinecartMeterListener extends VehicleListener {
                     plugin.resetDistanceCounter(player);
                     World world = l.getWorld();
                     long world_time = world.getTime();
-                    int world_hh = world_time/1000;
-                    int world_mm = (world_time - world_hh)*60/1000;
+                    long world_hh = world_time/1000;
+                    long world_mm = (world_time - world_hh)*60/1000;
                     System.out.println("[DEBUG]: full time of " + world.getName() + ": " + world.getFullTime());
                     System.out.println("[DEBUG]: time of " + world.getName() + ": " + world.getTime()+ " " + world_hh +":"+world_mm);
                     player.sendMessage("[DEBUG]: full time of " + world.getName() + ": " + world.getFullTime());
@@ -126,8 +126,8 @@ public class MinecartMeterListener extends VehicleListener {
                 System.out.println("[DEBUG]: player " + player.getName() + "have traveled " + plugin.getDistanceCounter(player) + " meters.");
                 World world = l.getWorld();
                 long world_time = world.getTime();
-                int world_hh = world_time/1000;
-                int world_mm = (world_time - world_hh)*60/1000;
+                long world_hh = world_time/1000;
+                long world_mm = (world_time - world_hh)*60/1000;
 
                 System.out.println("[DEBUG]: full time of " + world.getName() + ": " + world.getFullTime());
                 System.out.println("[DEBUG]: time of " + world.getName() + ": " + world.getTime()+ " " + world_hh +":"+world_mm);
